Implement logic class

+++ src/main/java/ru/tododecider/model/TaskDecider.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.tododecider.dao.TaskDao;
 
+import java.util.Collections;
 import java.util.PriorityQueue;
 import java.util.Queue;
 
@@ -26,33 +27,48 @@ public class TaskDecider {
 
     public Task getTopTask(){
         initAllTasks();
-        return allTasks.peek();
+        return allTasks.peek(); //todo поменять на элемент
     }
 
     public void addTask(Task t) {
-        allTasks.add(t);
-        taskDao.save(t);
+        if(t!=null) {
+            allTasks.add(t);
+            taskDao.save(t);
+        }
     }
 
     public void executeTopTask(){
-        Task t = allTasks.poll();
-        taskDao.delete(t.getName());
+        if(!noTasks()) {
+            Task t = allTasks.remove();
+            taskDao.delete(t.getName());
+        }
     }
 
     public void cancelTopTask(){
-        Task t = allTasks.poll();
-        taskDao.delete(t.getName());
+        if(!noTasks()) {
+            Task t = allTasks.remove();
+            taskDao.delete(t.getName());
+        }
     }
 
     public boolean noTasks(){
-        initAllTasks();
         return allTasks.isEmpty();
     }
 
     public void postponeTopTask(){
         if(!noTasks()) {
-            allTasks.add(allTasks.poll());
+            Task t = allTasks.remove();
+            t.setWaiting(0);
+            t.updatePriority();
+            incrementAllTasksWaiting();
+            allTasks.add(t);
         }
     }
 
+    public void incrementAllTasksWaiting(){
+        for (Task t:allTasks) {
+            t.setWaiting(t.getWaiting() + 1);
+            t.updatePriority();
+        }
+    }
 }
